Fix compiler warnings exposed by clang's '-Weverything'
* Add '-Weverything' to compilers flags if the compiler is
  clang and build type is 'Debug'.

* Fix compiler warnings exposed by -Weverything.

* Use pragmas to silence some warnings in murmur3 and
  one warning in octaspire_container_utf8_string,
  when using -Weverything with clang.

* Make the use of ANSI colors in greatest.h configurable
  and enable them if not in Windows.

+++ test/test_utf8.c
@@ -829,7 +829,7 @@ TEST octaspire_utf8_encode_and_then_decode_a_short_string_test(void)
         0x233B4
     };
 
-    int      expectedNumOctets[11] =
+    size_t expectedNumOctets[11] =
     {
         1,
         3,
@@ -862,7 +862,7 @@ TEST octaspire_utf8_encode_and_then_decode_a_short_string_test(void)
                 &numoctets));
 
         ASSERT_EQ(expectedResults[i],   result);
-        ASSERT_EQ(expectedNumOctets[i], numoctets);
+        ASSERT_EQ(((int)expectedNumOctets[i]), numoctets);
     }
 
     PASS();
